refactor: modularize docs knowledgebase page
